+++ lib/graphql/execution.rb
@@ -1,5 +1,6 @@
 require "graphql/execution/boxed"
 require "graphql/execution/directive_checks"
 require "graphql/execution/execute"
+require "graphql/execution/field_result"
 require "graphql/execution/selection_result"
 require "graphql/execution/typecast"
