+++ spec/graphql/execution/execute_spec.rb
@@ -1,0 +1,3 @@
+require "spec_helper"
+
+ExecuteSuite = GraphQL::Compatibility::ExecutionSpecification.build_suite(GraphQL::Execution::Execute)
