+++ spec/graphql/query/executor_spec.rb
@@ -127,7 +127,9 @@ describe GraphQL::Query::Executor do
           "data" => { "cow" => nil },
           "errors" => [
             {
-              "message" => "Cannot return null for non-nullable field Cow.cantBeNullButIs"
+              "message" => "Cannot return null for non-nullable field Cow.cantBeNullButIs",
+              "locations"=>[{"line"=>1, "column"=>28}],
+              "path"=>["cow", "cantBeNullButIs"],
             }
           ]
         }
